@@ -10,32 +10,25 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint"
+    "lint": "next lint"
   },
   "dependencies": {
-    "@react-three/drei": "^9.114.0",
-    "@react-three/fiber": "^8.15.11",
-    "next": "14.2.18",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "@react-three/drei": "^9.122.0",
+    "@react-three/fiber": "^9.5.0",
+    "next": "^15.1.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "three": "^0.169.0"
   },
-  "overrides": {
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
-  },
-  "resolutions": {
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
-  },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.1.0",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/node": "^22",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "@types/three": "^0.169.0",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "14.2.18",
+    "eslint": "^9",
+    "eslint-config-next": "^15.1.0",
     "tailwindcss": "^4",
     "typescript": "^5"
   }
